modif mainActivity>>MonsterPage
