+++ app/src/main/res/layout/activity_monsters.xml
@@ -6,4 +6,148 @@
     android:layout_height="match_parent"
     tools:context=".MonstersActivity">
 
+    <TextView
+        android:id="@+id/text_name"
+        android:layout_width="0dp"
+        android:layout_height="52dp"
+        android:background="@color/colorAccent"
+        android:text="Fire Lion"
+        android:textAlignment="center"
+        android:textSize="35dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/image_type"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="8dp"
+        android:layout_marginEnd="8dp"
+        android:layout_marginLeft="300dp"
+        android:layout_marginRight="8dp"
+        android:layout_marginStart="300dp"
+        android:layout_marginTop="8dp"
+        app:layout_constraintBottom_toBottomOf="@+id/text_name"
+        app:layout_constraintEnd_toEndOf="@+id/text_name"
+        app:layout_constraintHorizontal_bias="0.559"
+        app:layout_constraintStart_toStartOf="@+id/text_name"
+        app:layout_constraintTop_toTopOf="@+id/text_name"
+        app:srcCompat="@drawable/bte_fire" />
+
+    <ImageView
+        android:id="@+id/image_arene"
+        android:layout_width="396dp"
+        android:layout_height="340dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/text_name"
+        app:srcCompat="@drawable/fire_habitat_8"
+        android:alpha="0.5"/>
+
+    <ImageView
+        android:id="@+id/image_center"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="8dp"
+        android:layout_marginEnd="8dp"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        app:layout_constraintBottom_toBottomOf="@+id/image_arene"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="@+id/image_arene"
+        app:layout_constraintTop_toTopOf="@+id/image_arene"
+        app:srcCompat="@drawable/fire_lion_0" />
+
+    <ImageView
+        android:id="@+id/image_n0"
+        android:layout_width="61dp"
+        android:layout_height="63dp"
+        android:layout_marginEnd="8dp"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="260dp"
+        app:layout_constraintEnd_toStartOf="@+id/image_n1"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/image_arene"
+        app:srcCompat="@drawable/fire_lion_0" />
+
+    <ImageView
+        android:id="@+id/image_n1"
+        android:layout_width="89dp"
+        android:layout_height="82dp"
+        android:layout_marginLeft="8dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="240dp"
+        app:layout_constraintEnd_toStartOf="@+id/image_n2"
+        app:layout_constraintStart_toEndOf="@+id/image_n0"
+        app:layout_constraintTop_toTopOf="@+id/image_arene"
+        app:srcCompat="@drawable/fire_lion_1" />
+
+    <ImageView
+        android:id="@+id/image_n2"
+        android:layout_width="87dp"
+        android:layout_height="82dp"
+        android:layout_marginTop="240dp"
+        app:layout_constraintEnd_toStartOf="@+id/image_n3"
+        app:layout_constraintStart_toEndOf="@+id/image_n1"
+        app:layout_constraintTop_toTopOf="@+id/image_arene"
+        app:srcCompat="@drawable/fire_lion_2" />
+
+    <ImageView
+        android:id="@+id/image_n3"
+        android:layout_width="124dp"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="8dp"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="200dp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/image_n2"
+        app:layout_constraintTop_toTopOf="@+id/image_arene"
+        app:srcCompat="@drawable/fire_lion_3" />
+
+    <TextView
+        android:id="@+id/text_level"
+        android:layout_width="113dp"
+        android:layout_height="38dp"
+        android:layout_marginLeft="32dp"
+        android:layout_marginStart="32dp"
+        android:layout_marginTop="8dp"
+        android:background="@color/colorAccent"
+        android:text="LEVEL"
+        android:textAlignment="center"
+        android:textSize="30dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/image_n1" />
+
+    <EditText
+        android:id="@+id/edit_level"
+        android:layout_width="63dp"
+        android:layout_height="47dp"
+        android:layout_marginLeft="16dp"
+        android:layout_marginStart="16dp"
+        android:ems="10"
+        android:hint="0"
+        android:inputType="number"
+        android:textSize="25dp"
+        android:textAlignment="center"
+        app:layout_constraintStart_toEndOf="@+id/text_level"
+        app:layout_constraintTop_toBottomOf="@+id/image_n3" />
+
+    <Button
+        android:id="@+id/button_lvl"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginEnd="8dp"
+        android:layout_marginRight="8dp"
+        android:text="Change"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/edit_level"
+        app:layout_constraintTop_toBottomOf="@+id/image_n3" />
+
 </android.support.constraint.ConstraintLayout>
